pyproject.toml
[project]
name = "sd-hrm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "factory-boy>=3.3.3",
    "markdown>=3.8.2",
    "pillow>=11.3.0",
]
